+++ src/main/java/br/com/covid19/service/OccupationService.java
@@ -12,13 +12,16 @@ import java.util.Optional;
 @Service
 public class OccupationService {
 
+    public static Integer POST = 1;
+    public static Integer PUT = 2;
+
     @Autowired
     OccupationRepository occupationRepository;
 
     public List<Occupation> listOfOccupation(){
         List<Occupation> occupationList = occupationRepository.findAll();
         if (occupationList.isEmpty()){
-            throw new OccupationException();
+            throw new OccupationException("Não há profissões cadastradas no banco de dados.");
         }
         return occupationList;
     }
@@ -26,27 +29,44 @@ public class OccupationService {
     public Occupation findOccupation(Integer id){
         Optional<Occupation> optionalOccupation = occupationRepository.findById(id);
         if (optionalOccupation.isEmpty()){
-            throw new OccupationException();
+            throw new OccupationException("Profissão de id "+id+" não encontrada.");
         }
         return optionalOccupation.get();
     }
 
-    public Occupation registerNewOccupation(Occupation occupation){
-        if (occupation.getName() == null){
-            throw new OccupationException();
-        }else if(occupationRepository.save(occupation).getName() == null){
-            throw new OccupationException();
+    public Occupation registerNewOccupation(Occupation occupation, Integer requestMethod){
+        if (occupation == null){
+            throw new OccupationException("Profissão nula. Especifique uma profissão válida.");
+        }else if (occupation.getName() == null){
+            throw new OccupationException("Nome nulo. Especifique o nome da profissão.");
         }
-        return occupationRepository.save(occupation);
+        if (requestMethod == this.POST){
+            return occupationRepository.save(occupation);
+        }else if(requestMethod == this.PUT){
+            if (occupation.getId() == null){
+                throw new OccupationException("Id nulo. Especifique o id da profissão para realizar a edição.");
+            }
+            if (occupationRepository.findById(occupation.getId()).isEmpty()){
+                throw new OccupationException("Profissão com id "+occupation.getId()+" não existe.");
+            }else{
+                return occupationRepository.save(occupation);
+            }
+        }else {
+            throw new OccupationException("Método inválido.");
+        }
     }
 
     public Occupation deleteOccupation(Integer id){
-        Optional<Occupation> optionalOccupation = occupationRepository.findById(id);
-        if (optionalOccupation.isEmpty()){
-            throw new OccupationException();
+        if(id == null){
+            throw new OccupationException("Id nulo. Especifique o id da profissão.");
+        }else{
+            Optional<Occupation> optionalOccupation = occupationRepository.findById(id);
+            if (optionalOccupation.isEmpty()){
+                throw new OccupationException("Profissão de id "+id+" não encontrada.");
+            }
+            occupationRepository.deleteById(id);
+            return optionalOccupation.get();
         }
-        occupationRepository.deleteById(id);
-        return optionalOccupation.get();
     }
 
 }
